% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divViz-package.R
\docType{package}
\name{divViz package}
\alias{divViz}
\alias{divViz package}
\alias{divVizpackage,}
\alias{divViz-package}
\title{divViz: Visualisation tool for biodiversity data}
\description{
R package for Shiny visualisation tool of biodiversity data.
}
\details{
divViz: Visualisation tool for biodiversity data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RS-eco/divViz}
  \item Report bugs at \url{https://github.com/RS-eco/divViz/issues}
}

}
\author{
RS-eco
}
