% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divViz-package.R
\docType{data}
\name{landkreise}
\alias{landkreise}
\title{administrative Landkreis divisions of bavaria}
\format{
A \code{data.frame} with 96 observation and 15 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}} 
and has been combined with the districts data.
}
\usage{
data(landkreise)
}
\description{
GADM Landkreis divisions of Bavaria, Germany
}
\details{
Administrative Landkreis divisions of Bavaria, Germany derived from GADM data.
}
