% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divViz-package.R
\docType{data}
\name{tk_district}
\alias{tk_district}
\title{administrative subdivisions of bavaria per TK grid cell}
\format{
A \code{data.frame} with 613 observation and 2 variables.
}
\usage{
data(tk_district)
}
\description{
Administrative subdivisions per TK grid cell of Bavaria, Germany
}
\details{
Administrative subdivisions per TK grid cell of Bavaria, Germany.
}
