% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divViz-package.R
\docType{data}
\name{districts}
\alias{districts}
\title{administrative subdivisions of bavaria}
\format{
A \code{data.frame} with 7 observation and 18 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.ldbv.bayern.de/produkte/weitere/opendata.html}}
}
\usage{
data(districts)
}
\description{
GADM subdivisions of Bavaria, Germany
}
\details{
Administrative subdivisions of Bavaria, Germany derived from GADM data.
}
