% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divViz-package.R
\docType{package}
\name{divViz package}
\alias{divViz package}
\alias{divVizpackage}
\title{divViz: Visualisation tool for biodiversity data}
\description{
R package for Shiny visualisation tool of biodiversity data.
}
\details{
divViz: Visualisation tool for biodiversity data
}
\author{
RS-eco
}
\keyword{package}
